requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
requests==2.32.3
python-dateutil==2.9.0.post0
psycopg[binary,pool]==3.2.1
python-dotenv==1.0.1
gradio==4.44.0
